renamed utils.py to ModelPredictionUtils.py
